+++ dcp/dcp-crawler.rb
@@ -104,3 +104,17 @@ class DcpCrawler
   end
 
 end
+
+def execute(dcp_query)
+  start = DateTime.now
+  DcpLogger.log.info("Started at #{start}")
+  dcp_query.analyze
+  DcpLogger.log.info(("Total time: #{(DateTime.now.to_time - start.to_time)}"))
+  Kernel.exit(0)
+end
+
+if $PROGRAM_NAME == __FILE__
+  query_url = ARGV[0]
+  get_links = DcpCrawler.new(query_url, "#{__dir__}/config/dcp-config.yml")
+  execute(get_links)
+end
